# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ocean Blog API",
        "description": "Blogs API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Dev version"
        }
    ],
    "paths": {
...
        "/categories/{category}/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Отримати список постів у категорії",
                "operationId": "fb6918ab488db1415c964db7c9202cd1",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID категорії",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "posts": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Створити новий пост",
                "description": "Створює новий пост у категорії та повертає його у форматі JSON.",
                "operationId": "a4578702b63bacbe3926e7c686dda5af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост створено успішно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Показати пост",
                "description": "Повертає один пост за ID у категорії.",
                "operationId": "d99512970f542647fda93716e0a1d9bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно отримано пост",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Оновити пост",
                "description": "Оновлює існуючий пост за ID у категорії.",
                "operationId": "1ff0cadc31e4b79bde50ec8b5513dbdc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пост оновлено успішно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Видалити пост",
                "description": "Видаляє пост за ID у категорії.",
                "operationId": "d4fd127d9c73e609bb91f2ba3476fbd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Пост видалено успішно"
                    },
                    "404": {
                        "description": "Пост не знайдено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "My First Post"
                    },
                    "content": {
                        "type": "string",
                        "example": "This is the content of the post."
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-03T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-03T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "PostCreateRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "My First Post"
                    },
                    "content": {
                        "type": "string",
                        "example": "This is the content of the post."
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "PostUpdateRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Updated Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Updated content."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}